+++ src/pattern/construct/construct.spec.js
@@ -4,14 +4,14 @@ const {assert} = require('chai');
 
 describe('construct Tests', () => {
     it('can define null safe Value', () => {
-        assert.isNull(Create(undefined));
+        assert.isUndefined(Create(undefined, undefined));
         assert.equal(Create(undefined, 2), 2);
-        assert.isNull(Create(null));
-        assert.equal(Create(1), 1);
-        assert.equal(Create(false), false);
-        assert.equal(Create([1,2,3]).ElementAt(0), 1);
-        assert.equal(Create([1,2,3]).ElementAt(1), 2);
-        assert.equal(Create([1,2,3]).ElementAt(2), 3);
+        assert.isUndefined(Create(null, undefined));
+        assert.equal(Create(1, 0), 1);
+        assert.equal(Create(false, false), false);
+        assert.equal(Create([1,2,3], []).ElementAt(0), 1);
+        assert.equal(Create([1,2,3], []).ElementAt(1), 2);
+        assert.equal(Create([1,2,3], []).ElementAt(2), 3);
     });
 
     it('can create with factory method', () => {